refactor(dashboard): split coaching-sessions-card.tsx into focused sibling files
The single-file 874-line component is now five sibling files (~100-220
lines each), each with a single concern:

- coaching-sessions-card.tsx       — orchestrator + inline states
- coaching-sessions-card-header.tsx — title, chips, view toggle
- coaching-sessions-filters.tsx     — time-window enum/maps + popover
- coaching-sessions-list-view.tsx   — Tabs + master-detail layout
- coaching-sessions-row.tsx         — single SessionRow
- coaching-sessions-hover-detail.tsx — right-panel content

Public surface unchanged — CoachingSessionsCard is still the only export
the dashboard imports. No behavior changes; tsc + 887 unit tests + 32
chromium e2e tests all green.
